+++ Libraries/AssemblyLib/App.cs
@@ -1,6 +1,5 @@
 ﻿using System.Reflection;
 using System.Runtime.Loader;
-using AssemblyLib.DirectMapper;
 using AssemblyLib.Helpers;
 using AssemblyLib.Models;
 using Microsoft.Extensions.DependencyInjection;
