+++ Libraries/AssemblyLib/AttributeFactory/Builders/AsyncAttributeBuilder.cs
@@ -2,7 +2,7 @@ using AsmResolver.DotNet;
 using AsmResolver.DotNet.Signatures;
 using SPTarkov.DI.Annotations;
 
-namespace AssemblyLib.DirectMapper.AttributeFactory.Builders;
+namespace AssemblyLib.AttributeFactory.Builders;
 
 [Injectable]
 public class AsyncAttributeBuilder(ILogger<AsyncAttributeBuilder> logger, DataProvider dataProvider) : IAttributeBuilder
